requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
openai>=1.0.0
elevenlabs>=0.2.0
moviepy>=1.0.3
pillow>=10.0.0
requests>=2.31.0 